# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: HoldCo Unified API
  version: 1.0.0
  description: Unified API for HoldCo subsidiaries (wholesale, retail, reseller credit, digital commerce, logistics).
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging.api.example.com
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /v1/finance/accounts:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Finance]
      operationId: listAccounts
      summary: List chart of accounts
      x-rbac:
        permissions: [finance.chart_of_accounts.manage]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartOfAccountListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Finance]
      operationId: createAccount
      summary: Create chart of account
      x-rbac:
        permissions: [finance.chart_of_accounts.manage]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChartOfAccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChartOfAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/cost-centers:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: listCostCenters
      summary: List cost centers
      x-rbac:
        permissions: [finance.cost_centers.manage]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCenterListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Finance]
      operationId: createCostCenter
      summary: Create cost center
      x-rbac:
        permissions: [finance.cost_centers.manage]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCostCenterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCenter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/fiscal-periods:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    get:
      tags: [Finance]
      operationId: listFiscalPeriods
      summary: List fiscal periods
      x-rbac:
        permissions: [finance.fiscal_periods.manage]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPeriodListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Finance]
      operationId: createFiscalPeriod
      summary: Create fiscal period
      x-rbac:
        permissions: [finance.fiscal_periods.manage]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFiscalPeriodRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiscalPeriod'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/journal-entries:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: listJournalEntries
      summary: List journal entries
      x-rbac:
        permissions: [finance.journal_entries.manage]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Finance]
      operationId: createJournalEntry
      summary: Create journal entry
      x-rbac:
        permissions: [finance.journal_entries.manage]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalEntryRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntry'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/exports/journals:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: exportJournalEntries
      summary: Export journal entries
      x-rbac:
        permissions: [finance.exports.read]
      parameters:
        - $ref: '#/components/parameters/ExportFormat'
        - $ref: '#/components/parameters/FiscalPeriodId'
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceExportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/exports/invoices:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: exportInvoices
      summary: Export invoices
      x-rbac:
        permissions: [finance.exports.read]
      parameters:
        - $ref: '#/components/parameters/ExportFormat'
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/InvoiceType'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceExportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/exports/credit-notes:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: exportCreditNotes
      summary: Export credit notes
      x-rbac:
        permissions: [finance.exports.read]
      parameters:
        - $ref: '#/components/parameters/ExportFormat'
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/InvoiceType'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceExportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/exports/intercompany:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: exportIntercompanyInvoices
      summary: Export intercompany invoices
      x-rbac:
        permissions: [finance.exports.read]
      parameters:
        - $ref: '#/components/parameters/ExportFormat'
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceExportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/exports/payments:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - $ref: '#/components/parameters/SubsidiaryId'
    get:
      tags: [Finance]
      operationId: exportPayments
      summary: Export payments
      x-rbac:
        permissions: [finance.exports.read]
      parameters:
        - $ref: '#/components/parameters/ExportFormat'
        - $ref: '#/components/parameters/Period'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceExportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/month-close:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: runMonthClose
      summary: Run month close workflow
      x-rbac:
        permissions: [finance.month_close.run]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonthCloseRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthCloseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/cost-pools:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: createCostPool
      summary: Create or update a cost pool
      x-rbac:
        permissions: [finance.cost_pools.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCostPoolRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostPoolResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/cost-pools/{id}/allocate:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Finance]
      operationId: allocateCostPool
      summary: Allocate cost pool
      x-rbac:
        permissions: [finance.cost_pools.manage]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocateCostPoolResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/intercompany/{period}/generate:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - name: period
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Finance]
      operationId: generateIntercompanyInvoices
      summary: Generate intercompany invoices for a period
      x-rbac:
        permissions: [finance.intercompany.generate]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateIntercompanyInvoicesRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntercompanyInvoicesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/invoices/{id}/issue:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Finance]
      operationId: issueInvoice
      summary: Issue an intercompany invoice
      x-rbac:
        permissions: [finance.invoices.issue]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinanceInvoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/ledger/post-period:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: postLedgerForPeriod
      summary: Post all invoices for a period to the ledger
      x-rbac:
        permissions: [finance.ledger.post]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostLedgerPeriodRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostLedgerPeriodResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/payments/intercompany:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: recordIntercompanyPayment
      summary: Record intercompany payment
      x-rbac:
        permissions: [finance.payments.record]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordIntercompanyPaymentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordIntercompanyPaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/wht-schedule/{issuerCompanyId}/{period}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - name: issuerCompanyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: period
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Finance]
      operationId: getWhtSchedule
      summary: Get WHT remittance schedule
      x-rbac:
        permissions: [finance.wht.schedule.read]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhtScheduleResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/wht-remit:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: remitWhtCredits
      summary: Mark WHT credits as remitted
      x-rbac:
        permissions: [finance.wht.remit]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhtRemitRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhtRemitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/vat/{companyId}/{period}:
    parameters:
      - $ref: '#/components/parameters/GroupId'
      - name: companyId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: period
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Finance]
      operationId: generateVatReturn
      summary: Generate VAT return for a period
      x-rbac:
        permissions: [finance.vat.generate]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VatReturn'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/vat/file:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: fileVatReturn
      summary: File VAT return for a period
      x-rbac:
        permissions: [finance.vat.file]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileVatReturnRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VatReturn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/finance/period-lock/lock:
    parameters:
      - $ref: '#/components/parameters/GroupId'
    post:
      tags: [Finance]
      operationId: lockPeriod
      summary: Lock a period
      x-rbac:
        permissions: [finance.period_lock.manage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeriodLockRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeriodLock'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  parameters:
    GroupId:
      name: X-Group-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant group identifier
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for safe retries of write operations
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
...
  schemas:
...
    CreateChartOfAccountRequest:
      type: object
      required: [code, name, type]
      properties:
        code:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/ChartOfAccountType'
        parent_id:
          type: string
          format: uuid
      example:
        code: "4000"
        name: Sales Revenue
        type: income
    ChartOfAccountListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ChartOfAccount'
        meta:
          $ref: '#/components/schemas/PageMeta'
      example:
        data:
          - id: 2b7b8d2a-1f15-4b7c-9c7f-5f65d8b2f9ac
            code: "4000"
            name: Sales Revenue
            type: income
        meta:
          limit: 50
          offset: 0
          total: 1
...